CB-40 refactor: change enroll flow
등록 플로우 순서 변경
- Step1. 이름
- Step2. 사진
- Step3. 견종
- Step4. 나이
- Step5. 상세 정보(성별, 몸무게, 활동레벨, 중성화, 임신/수유여부)
