+++ src/pages/RegisterPet/Step1.tsx
@@ -1,80 +1,46 @@
-/* eslint-disable no-unused-vars */
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
 import { useForm } from 'react-hook-form';
-import FormButton from '@/components/Form/FormButton';
 import FormInput from '@/components/Form/FormInput';
-import { concatClasses } from '@/utils/libs/concatClasses';
+import FormButton from '@/components/Form/FormButton';
+import { useEffect } from 'react';
+import { useAppDispatch, useAppSelector } from '@/store/config';
 import { selectRegisterInfo, setRegisterInfo } from '@/store/slices/registerPetSlice';
-import { PetSexType } from '@/@type/pet';
-import useSelectImage from '@/utils/hooks/useSelectImage';
-import { ButtonWrapper, PageContainer, QuestionText, Form, QuestionWrapper } from './index.style';
-import { INextStep } from './type';
+import { ButtonWrapper, Form, PageContainer, QuestionText, SubQuestionText } from './index.style';
 
-interface IStep1Form {
+interface Step1Form {
   name: string;
-  sex: PetSexType;
-  isSpayed: boolean;
-  isPregnant: boolean;
 }
-export default function Step1({ goNextStep }: INextStep) {
-  const dispatch = useDispatch();
-  const registerInfo = useSelector(selectRegisterInfo);
-  const { imageFile: petImage, handleChangeImage } = useSelectImage({
-    initImageFile: registerInfo.petImage,
-  });
+export default function Step1({ goNextStep }: any) {
+  const dispatch = useAppDispatch();
+  const registerInfo = useAppSelector(selectRegisterInfo);
   const {
     register,
     handleSubmit,
-    watch,
-    setValue,
     formState: { errors },
-  } = useForm<IStep1Form>();
+    setValue,
+    watch,
+  } = useForm<Step1Form>();
 
-  const onValidSubmit = (submitData: IStep1Form) => {
-    const { name, sex, isSpayed, isPregnant } = submitData;
-
-    dispatch(setRegisterInfo({ petImage, name, sex, isSpayed, isPregnant }));
+  const isButtonDisabled = !watch('name');
+  const onSubmitForm = ({ name }: Step1Form) => {
+    dispatch(setRegisterInfo({ name }));
     goNextStep();
   };
 
   useEffect(() => {
     if (!registerInfo) return;
-    console.log(registerInfo);
     setValue('name', registerInfo.name);
-    setValue('sex', registerInfo.sex);
-    setValue('isSpayed', registerInfo.isSpayed);
-    setValue('isPregnant', registerInfo.isPregnant);
   }, []);
 
   return (
     <PageContainer>
-      <div>
-        <QuestionText>함께하는 아이에 대해 알고싶어요!</QuestionText>
-      </div>
-      <Form onSubmit={handleSubmit(onValidSubmit)}>
-        <div className="flex flex-col gap-6">
-          <div className="flex flex-col items-center justify-center">
-            <div className="relative felx items-center justify-center">
-              <label
-                className="absolute bottom-2 right-2 h-8 w-8 bg-primary-900 rounded-md"
-                htmlFor="pet-thumbnail"
-              ></label>
-              <img
-                alt=""
-                src={petImage}
-                className="bg-gray-200 h-32 w-32 rounded-full overflow-hidden"
-              />
-            </div>
-            <input
-              type="file"
-              id="pet-thumbnail"
-              accept="image/jpg, image/png, image/jpeg"
-              onChange={handleChangeImage}
-            />
-          </div>
+      <div className="flex flex-col gap-5 h-full">
+        <div>
+          <QuestionText>이름이 어떻게 되나요?</QuestionText>
+          <SubQuestionText>되도록 10자 이내였으면 좋겠어요!</SubQuestionText>
+        </div>
+        <Form onSubmit={handleSubmit(onSubmitForm)}>
           <FormInput
-            label="이름"
+            label=""
             name="pet-name"
             placeholder="반려동물의 이름을 입력해주세요"
             rules={register('name', {
@@ -85,62 +51,11 @@ export default function Step1({ goNextStep }: INextStep) {
             isError={!!errors.name?.message}
             errorMessage={errors.name?.message}
           />
-          <div>
-            <div className="flex gap-1 items-center mb-2">
-              <div className="flex-1 text-center w-1/2">
-                <input
-                  type="radio"
-                  className="hidden"
-                  value="MALE"
-                  id="pet-sex-man"
-                  {...register('sex', { required: '성별을 선택해주세요' })}
-                />
-                <label
-                  htmlFor="pet-sex-man"
-                  className={concatClasses(
-                    'border border-primary-900 rounded-md w-full block',
-                    watch('sex') === 'MALE' ? 'bg-primary-100 text-primary-900' : '',
-                  )}
-                >
-                  남자
-                </label>
-              </div>
-              <div className="flex-1 text-center w-1/2">
-                <input
-                  type="radio"
-                  className="hidden"
-                  value="FEMALE"
-                  id="pet-sex-woman"
-                  {...register('sex', { required: '성별을 선택해주세요' })}
-                />
-                <label
-                  htmlFor="pet-sex-woman"
-                  className={concatClasses(
-                    'border border-primary-900 rounded-md w-full block',
-                    watch('sex') === 'FEMALE' ? 'bg-primary-100 text-primary-900' : '',
-                  )}
-                >
-                  여자
-                </label>
-              </div>
-            </div>
-            <p className="text-primary-900 text-sm">{errors.sex?.message}</p>
-          </div>
-          <div>
-            <QuestionWrapper>
-              <input type="checkbox" id="spayed" {...register('isSpayed')} />
-              <label htmlFor="spayed">중성화를 했나요?</label>
-            </QuestionWrapper>
-            <QuestionWrapper>
-              <input type="checkbox" id="pregnant" {...register('isPregnant')} />
-              <label htmlFor="pregnant">임신/수유 중인가요?</label>
-            </QuestionWrapper>
-          </div>
-        </div>
-        <ButtonWrapper>
-          <FormButton name="다음으로" disabled={false} />
-        </ButtonWrapper>
-      </Form>
+          <ButtonWrapper>
+            <FormButton name="다음으로" disabled={isButtonDisabled} />
+          </ButtonWrapper>
+        </Form>
+      </div>
     </PageContainer>
   );
 }
